refactor(pipelines): moved stage functions into separate files
