+++ vars/initAwsRegionParam.groovy
@@ -1,6 +1,6 @@
 #!groovy
 
-def String initAwsRegionParam(String regionStr, String builder_region){
+def call(String regionStr, String builder_region){
     if (regionStr == "random"){
         return builder_region
     }
